Various.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/math/Conversion.java
@@ -1,7 +1,7 @@
 package org.firstinspires.ftc.teamcode.util.math;
 
 public class Conversion {
-    public static double millsToSec(int mills) {
-        return mills / 1000.0;
+    public static double millsToSec(double mills) {
+        return mills / 1000;
     }
 }
